refactor(settings-share): split flagged functions under 150 lines
Refactor SettingsClient and ShareVideoPageClient
- settings: extract useSettings/useProfileForm/usePhotoUpload/useLoadUser
  hooks, a settingsHelpers util, and ProfileTab/AccountTab/SettingsTabs/
  AvatarSection/PhotoUploadCard/ProfileFormFields/ProfileFormActions
  presentational components.
- share: extract useViewTracking hook, getPreviewStage util, and
  ShareHeader/ShareSignupCTA components.
