+++ app/(signed)/settings/SettingsClient.tsx
@@ -1,616 +1,26 @@
 "use client";
-import React, { useEffect, useMemo, useRef, useState, ChangeEvent, FormEvent } from "react";
-import { useSession, signOut } from "next-auth/react";
-import { useRouter } from "next/navigation";
 import { ToastContainer } from "react-toastify";
 import "react-toastify/dist/ReactToastify.css";
-import { toast } from "sonner";
-import Image from "next/image";
 
-import { TABS } from "../../lib/constants";
+import { useSettings } from "./hooks/useSettings";
+import SettingsTabs from "./components/SettingsTabs";
+import ProfileTab from "./components/ProfileTab";
+import AccountTab from "./components/AccountTab";
 
 export const metadata = {
   titleText: "Analytics",
   iconSRC: "/majesticons_analytics.png",
 };
 const SettingsPage = () => {
-  const router = useRouter();
-  const { data: session } = useSession();
-  const [activeTab, setActiveTab] = useState("Profile");
-  const [form, setForm] = useState({
-    firstName: session?.user?.name?.split(" ")[0] || "",
-    lastName: session?.user?.name?.split(" ").slice(1).join(" ") || "",
-    email: session?.user?.email || "",
-    bio: "",
-    location: "",
-    website: "",
-    timezone: "",
-    image: "",
-  });
-  const [avatar, setAvatar] = useState("");
-  const fileInputRef = useRef<HTMLInputElement>(null);
-  const [photoFile, setPhotoFile] = useState<File | null>(null);
-  const [showPhotoCard, setShowPhotoCard] = useState(false);
-  const [isDragging, setIsDragging] = useState(false);
-
-  const [isUploading, setIsUploading] = useState(false);
-  const [isDirty, setIsDirty] = useState(false);
-  const [isSaving, setIsSaving] = useState(false);
-  const [isSendingPasswordReset, setIsSendingPasswordReset] = useState(false);
-  const [originalForm, setOriginalForm] = useState({ ...form });
-
-  const initials = useMemo(() => {
-    if (session?.user?.name) {
-      return session.user.name
-        .split(" ")
-        .map((part) => part[0])
-        .join("")
-        .toUpperCase()
-        .slice(0, 2);
-    }
-    return session?.user?.email?.[0].toUpperCase() || "U";
-  }, [session?.user]);
-
-  useEffect(() => {
-    setForm((prev) => ({
-      ...prev,
-      firstName: session?.user?.name?.split(" ")[0] || "",
-      lastName: session?.user?.name?.split(" ").slice(1).join(" ") || "",
-      email: session?.user?.email || "",
-    }));
-
-    const fetchUser = async () => {
-      const res = await fetch("/api/user/get");
-      const data = await res.json();
-
-      if (data.user) {
-        const user = data.user;
-        console.log(user.image, user.name);
-        setAvatar(user.image || "");
-
-        setForm((prev) => ({
-          ...prev,
-          bio: user.bio || "",
-          location: user.location || "",
-          website: user.website || "",
-          timezone: user.timezone || "",
-          image: user.image || "",
-          firstName: user.name?.split(" ")[0] || "",
-          lastName: user.name?.split(" ").slice(1).join(" ") || "",
-          email: user.email || "",
-        }));
-      }
-    };
-
-    fetchUser();
-  }, [session]);
-
-  useEffect(() => {
-    const fetchUser = async () => {
-      const res = await fetch("/api/user/get");
-      const data = await res.json();
-      if (data.user) {
-        const user = data.user;
-        const userFormData = {
-          firstName: user.name?.split(" ")[0] || "",
-          lastName: user.name?.split(" ").slice(1).join(" ") || "",
-          email: user.email || "",
-          bio: user.bio || "",
-          location: user.location || "",
-          website: user.website || "",
-          timezone: user.timezone || "",
-          image: user.image || "",
-        };
-        setForm(userFormData);
-        setOriginalForm(userFormData);
-      }
-    };
-    fetchUser();
-  }, [session]);
-
-  const hasChanges = useMemo(() => {
-    return JSON.stringify(form) !== JSON.stringify(originalForm);
-  }, [form, originalForm]);
-
-  const handleInputChange = (e: ChangeEvent<HTMLInputElement>) => {
-    setIsDirty(true);
-    setForm({ ...form, [e.target.name]: e.target.value });
-  };
-
-  const handlePhotoChange = (e: ChangeEvent<HTMLInputElement>) => {
-    if (e.target.files && e.target.files[0]) {
-      const file = e.target.files[0];
-      if (!file.type.startsWith("image/")) {
-        toast.error("Please upload an image file");
-        return;
-      }
-      setPhotoFile(file);
-      setIsDirty(true);
-      const reader = new FileReader();
-      reader.onload = (event) => {
-        if (event.target?.result) {
-          setAvatar(event.target.result as string);
-          setForm((prev) => ({
-            ...prev,
-            image: event.target?.result as string,
-          }));
-        }
-      };
-      reader.readAsDataURL(file);
-    }
-  };
-
-  const handleRemovePhoto = () => {
-    console.log("Remove photo clicked");
-    setPhotoFile(null);
-    setAvatar("");
-    setForm((prev) => ({ ...prev, image: "" }));
-    setIsDirty(true);
-    if (fileInputRef.current) {
-      fileInputRef.current.value = "";
-    }
-  };
-
-  const handleSave = async (e: FormEvent<HTMLFormElement>) => {
-    e.preventDefault();
-    setIsSaving(true);
-    try {
-      let imageUrl = form.image;
-
-      if (photoFile) {
-        setIsUploading(true);
-        const formData = new FormData();
-        formData.append("file", photoFile);
-        formData.append("upload_preset", process.env.NEXT_PUBLIC_CLOUDINARY_UPLOAD_PRESET!);
-
-        const uploadRes = await fetch("/api/upload", {
-          method: "POST",
-          body: formData,
-        });
-        const uploadData = await uploadRes.json();
-
-        console.log("Upload Status:", uploadRes.status);
-        console.log("Upload Data:", uploadData);
-
-        if (!uploadRes.ok) {
-          console.error("Upload failed with status:", uploadRes.status);
-          toast.error(uploadData.error || "Failed to upload photo");
-          setIsUploading(false);
-          setIsSaving(false);
-          return;
-        }
-
-        if (!uploadData.secure_url) {
-          console.error("No secure_url in response:", uploadData);
-          toast.error(uploadData.error || "Failed to get image URL from upload");
-          setIsUploading(false);
-          setIsSaving(false);
-          return;
-        }
-
-        imageUrl = uploadData.secure_url;
-        console.log("Image URL after upload:", imageUrl);
-        setIsUploading(false);
-      }
-
-      console.log("Saving with image URL:", imageUrl);
-
-      const res = await fetch("/api/user/update", {
-        method: "POST",
-        headers: { "Content-Type": "application/json" },
-        body: JSON.stringify({ ...form, image: imageUrl }),
-      });
-
-      const data = await res.json();
-      console.log("Save response:", data);
-
-      if (res.ok) {
-        await new Promise((resolve) => setTimeout(resolve, 500));
-        router.refresh();
-        toast.success("Changes updated successfully!");
-        setIsDirty(false);
-        setOriginalForm({ ...form, image: imageUrl });
-        setPhotoFile(null);
-        setAvatar(imageUrl || "");
-        console.log("Dispatching photoUpdated event");
-        window.dispatchEvent(new Event("photoUpdated"));
-        if (fileInputRef.current) {
-          fileInputRef.current.value = "";
-        }
-      } else {
-        toast(`Update failed: ${data.error}`);
-      }
-    } catch {
-      toast("Failed to save changes");
-    } finally {
-      setIsSaving(false);
-      setIsUploading(false);
-    }
-  };
-
-  const handleCancel = async () => {
-    const res = await fetch("/api/user/get");
-    const data = await res.json();
-    if (data.user) {
-      const user = data.user;
-      setForm({
-        firstName: user.name?.split(" ")[0] || "",
-        lastName: user.name?.split(" ").slice(1).join(" ") || "",
-        email: user.email || "",
-        bio: user.bio || "",
-        location: user.location || "",
-        website: user.website || "",
-        timezone: user.timezone || "",
-        image: user.image || "",
-      });
-      setAvatar(user.image || "");
-      setPhotoFile(null);
-      setIsDirty(false);
-      if (fileInputRef.current) {
-        fileInputRef.current.value = "";
-      }
-    }
-  };
-
-  const handleDeleteAccount = async () => {
-    const confirmed = confirm(
-      "Are you sure you want to delete your account? This action cannot be undone."
-    );
-    if (!confirmed) {
-      return;
-    }
-
-    const res = await fetch("/api/user/delete", {
-      method: "DELETE",
-    });
-
-    const data = await res.json();
-
-    if (res.ok) {
-      alert("Your account has been deleted.");
-      signOut({ callbackUrl: "/" });
-    } else {
-      alert(`Failed to delete account: ${data.error}`);
-    }
-  };
-
-  const handleSendPasswordReset = async () => {
-    try {
-      setIsSendingPasswordReset(true);
-      const res = await fetch("/api/auth/request-reset", {
-        method: "POST",
-        headers: { "Content-Type": "application/json" },
-        body: JSON.stringify({ authenticatedOnly: true }),
-      });
-      const data = await res.json();
-      if (!res.ok) {
-        toast.error(data?.error || "Failed to send reset email.");
-        return;
-      }
-      toast.success("Password reset link sent to your email.");
-    } catch (error) {
-      console.error("Failed to send password reset:", error);
-      toast.error("Failed to send reset email.");
-    } finally {
-      setIsSendingPasswordReset(false);
-    }
-  };
-
-  const handleDragOver = (e: React.DragEvent<HTMLDivElement>) => {
-    e.preventDefault();
-    e.stopPropagation();
-    setIsDragging(true);
-  };
-
-  const handleDragLeave = (e: React.DragEvent<HTMLDivElement>) => {
-    e.preventDefault();
-    e.stopPropagation();
-    if (e.currentTarget === e.target) {
-      setIsDragging(false);
-    }
-  };
-
-  const handleDrop = (e: React.DragEvent<HTMLDivElement>) => {
-    e.preventDefault();
-    e.stopPropagation();
-    setIsDragging(false);
-
-    const files = e.dataTransfer.files;
-    if (files && files[0]) {
-      const file = files[0];
-      if (!file.type.startsWith("image/")) {
-        toast.error("Please upload an image file");
-        return;
-      }
-      setPhotoFile(file);
-      setIsDirty(true);
-      setShowPhotoCard(false);
-
-      const reader = new FileReader();
-      reader.onload = (event) => {
-        if (event.target?.result) {
-          setAvatar(event.target.result as string);
-          setForm((prev) => ({
-            ...prev,
-            image: event.target?.result as string,
-          }));
-        }
-      };
-      reader.readAsDataURL(file);
-    }
-  };
+  const settings = useSettings();
 
   return (
     <div className="settings-page min-h-screen bg-[#F3F0FC]">
-      <div className="tabs flex flex-wrap items-center gap-2 px-2 sm:px-4 md:px-8 pb-3 pt-4 bg-white border-b border-gray-200 overflow-x-auto">
-        {TABS.map((tab) => (
-          <button
-            key={tab}
-            className={`tab flex-1 min-w-[120px] px-3 sm:px-6 py-2 border rounded-lg text-base sm:text-lg font-medium transition-colors focus:outline-none whitespace-nowrap ${
-              activeTab === tab
-                ? "active bg-[#7C5CFC] text-white shadow"
-                : "bg-transparent text-gray-500 hover:bg-[#ede7fa]"
-            }`}
-            onClick={() => setActiveTab(tab)}
-          >
-            {tab}
-          </button>
-        ))}
-      </div>
+      <SettingsTabs activeTab={settings.activeTab} setActiveTab={settings.setActiveTab} />
 
-      {activeTab === "Profile" && (
-        <div className="px-2 sm:px-4 md:px-8 lg:px-16 xl:px-24">
-          <div className="w-full mx-auto mt-8 mb-2">
-            <h2 className="text-2xl font-bold mb-1">Profile Information</h2>
-            <p className="text-gray-500 mb-6">Manage your profile settings here.</p>
-          </div>
-          <form
-            className="card w-full mx-auto mt-2 mb-12 bg-white rounded-xl border border-[#ede7fa] shadow-none p-4 sm:p-6 md:p-8 lg:p-10 relative"
-            onSubmit={handleSave}
-            noValidate
-          >
-            <div className="flex flex-col sm:flex-row items-center gap-4 sm:gap-6 mb-8">
-              <div className="w-24 h-24">
-                <div className="avatar-placeholder w-24 h-24 rounded-full bg-[#F3F0FC] flex items-center justify-center text-3xl font-bold text-[#7C5CFC] border-2 border-[#E0D7FF] cursor-pointer hover:opacity-80 transition-opacity">
-                  {avatar && avatar.trim() ? (
-                    <Image
-                      key={avatar}
-                      src={avatar}
-                      alt="Avatar"
-                      width={96}
-                      height={96}
-                      className="w-full h-full object-cover rounded-full"
-                      unoptimized
-                    />
-                  ) : (
-                    initials
-                  )}
-                </div>
-              </div>
-              <div className="flex flex-col sm:flex-row gap-2 sm:gap-4 w-full sm:w-auto">
-                <button
-                  type="button"
-                  disabled={isSaving || isUploading}
-                  className="px-5 py-2 rounded-lg bg-[#7C5CFC] text-white font-semibold shadow hover:bg-[#8A76FC] transition w-full sm:w-auto disabled:opacity-50 disabled:cursor-not-allowed"
-                  onClick={() => setShowPhotoCard(true)}
-                >
-                  Change Photo
-                </button>
-                {avatar && avatar.trim() && (
-                  <button
-                    type="button"
-                    disabled={isSaving || isUploading}
-                    className="remove-photo-btn px-5 py-2 rounded-lg bg-white border border-gray-200 text-[#8A76FC] font-semibold shadow hover:bg-[#F3F0FC] transition w-full sm:w-auto disabled:opacity-50 disabled:cursor-not-allowed flex items-center justify-center gap-2"
-                    onClick={handleRemovePhoto}
-                  >
-                    <Image src="/icons/si_bin-line.png" alt="Remove" width={18} height={18} />
-                    Remove Photo
-                  </button>
-                )}
-                <input
-                  type="file"
-                  accept="image/*"
-                  ref={fileInputRef}
-                  className="hidden"
-                  onChange={handlePhotoChange}
-                />
-              </div>
-            </div>
+      {settings.activeTab === "Profile" && <ProfileTab settings={settings} />}
 
-            {showPhotoCard && (
-              <div
-                className="fixed inset-0 backdrop-blur-lg bg-white/15 flex items-center justify-center z-50 p-4"
-                onDragOver={handleDragOver}
-                onDragLeave={handleDragLeave}
-                onDrop={handleDrop}
-                onClick={(e) => {
-                  if (e.target === e.currentTarget) {
-                    setShowPhotoCard(false);
-                  }
-                }}
-              >
-                <div
-                  className={`bg-[#F3F0FC] rounded-3xl max-w-md w-full p-12 border-2 transition-all ${
-                    isDragging
-                      ? "border-solid border-[#7C5CFC] bg-[#FFFBFE] shadow-lg"
-                      : "border-dashed border-[#A594F9]"
-                  }`}
-                >
-                  <div className="flex flex-col items-center pointer-events-none">
-                    <Image
-                      src="/icons/zondicons_upload.png"
-                      alt="Upload"
-                      width={56}
-                      height={56}
-                      className={`mb-6 transition-transform ${isDragging ? "scale-110" : ""}`}
-                    />
-
-                    <h3 className="text-lg font-semibold text-gray-800 mb-3 text-center">
-                      {isDragging ? "Drop your image here" : "Drag and Drop files to upload"}
-                    </h3>
-
-                    <p className="text-gray-500 text-center mb-5">or</p>
-                  </div>
-
-                  <button
-                    type="button"
-                    onClick={() => {
-                      setShowPhotoCard(false);
-                      fileInputRef.current?.click();
-                    }}
-                    className="w-full px-8 py-2.5 bg-[#7C5CFC] text-white font-semibold rounded-lg hover:bg-[#8A76FC] transition mb-4"
-                  >
-                    Browse
-                  </button>
-
-                  <p className="text-gray-500 text-xs text-center">
-                    Supported files: JPEG, PNG, GIF
-                  </p>
-                  <p className="text-red-500 text-xs text-center">Max file size: 3MB</p>
-                </div>
-              </div>
-            )}
-
-            <div className="grid grid-cols-1 sm:grid-cols-2 gap-4 sm:gap-6 mb-6">
-              <div>
-                <label className="block text-gray-600 mb-2">First Name</label>
-                <input
-                  type="text"
-                  name="firstName"
-                  value={form.firstName}
-                  onChange={handleInputChange}
-                  placeholder="Enter your first name here"
-                  className="w-full px-4 py-3 rounded-lg border border-gray-200 bg-[#F8F6FF] text-gray-700 focus:outline-none focus:ring-2 focus:ring-[#A594F9] shadow-sm"
-                />
-              </div>
-              <div>
-                <label className="block text-gray-600 mb-2">Last Name</label>
-                <input
-                  type="text"
-                  name="lastName"
-                  value={form.lastName}
-                  onChange={handleInputChange}
-                  placeholder="Enter your last name here"
-                  className="w-full px-4 py-3 rounded-lg border border-gray-200 bg-[#F8F6FF] text-gray-700 focus:outline-none focus:ring-2 focus:ring-[#A594F9] shadow-sm"
-                />
-              </div>
-              <div className="sm:col-span-2">
-                <label className="block text-gray-600 mb-2">Email address</label>
-                <input
-                  type="email"
-                  name="email"
-                  value={form.email}
-                  onChange={handleInputChange}
-                  placeholder="Enter your email id here"
-                  className="w-full px-4 py-3 rounded-lg border border-gray-200 bg-[#F8F6FF] text-gray-700 focus:outline-none focus:ring-2 focus:ring-[#A594F9] shadow-sm"
-                  disabled
-                />
-              </div>
-              <div className="sm:col-span-2">
-                <label className="block text-gray-600 mb-2">Bio</label>
-                <input
-                  type="text"
-                  name="bio"
-                  value={form.bio}
-                  onChange={handleInputChange}
-                  placeholder="Tell us about yourself"
-                  className="w-full px-4 py-3 rounded-lg border border-gray-200 bg-[#F8F6FF] text-gray-700 focus:outline-none focus:ring-2 focus:ring-[#A594F9] shadow-sm"
-                />
-              </div>
-              <div>
-                <label className="block text-gray-600 mb-2">Location</label>
-                <input
-                  type="text"
-                  name="location"
-                  value={form.location}
-                  onChange={handleInputChange}
-                  placeholder="City, Country"
-                  className="w-full px-4 py-3 rounded-lg border border-gray-200 bg-[#F8F6FF] text-gray-700 focus:outline-none focus:ring-2 focus:ring-[#A594F9] shadow-sm"
-                />
-              </div>
-              <div>
-                <label className="block text-gray-600 mb-2">Website</label>
-                <input
-                  type="text"
-                  name="website"
-                  value={form.website}
-                  onChange={handleInputChange}
-                  placeholder="Paste your link here"
-                  className="w-full px-4 py-3 rounded-lg border border-gray-200 bg-[#F8F6FF] text-gray-700 focus:outline-none focus:ring-2 focus:ring-[#A594F9] shadow-sm"
-                />
-              </div>
-            </div>
-            {isDirty && hasChanges && (
-              <div className="flex flex-col sm:flex-row justify-end gap-2 sm:gap-4 mt-8">
-                <button
-                  type="button"
-                  className="px-8 py-3 cursor-pointer rounded-lg bg-white border border-gray-200 text-[#7C5CFC] font-semibold shadow hover:bg-[#ede7fa] transition w-full sm:w-auto"
-                  onClick={handleCancel}
-                >
-                  Cancel
-                </button>
-                <button
-                  type="submit"
-                  disabled={isSaving || isUploading}
-                  className="px-8 py-3 cursor-pointer rounded-lg bg-[#7C5CFC] text-white font-semibold shadow hover:bg-[#8A76FC] transition w-full sm:w-auto disabled:opacity-50 disabled:cursor-not-allowed"
-                >
-                  {isSaving ? "Saving..." : "Save Changes"}
-                </button>
-              </div>
-            )}
-          </form>
-        </div>
-      )}
-
-      {activeTab === "Account" && (
-        <div className="px-2 sm:px-4 md:px-8 lg:px-16 xl:px-24 flex flex-col">
-          <div className="w-full mb-12">
-            <h2 className="text-xl sm:text-2xl font-bold mb-4">Data Management</h2>
-            <div className="flex flex-col gap-3 sm:gap-4 px-4 sm:px-6">
-              <div className="card account-card flex flex-col sm:flex-row sm:items-center sm:justify-between bg-[#F3F0FC] rounded-lg border border-[#ede7fa] p-4 sm:px-6 sm:py-4 gap-3 sm:gap-4">
-                <div className="flex-1 min-w-0">
-                  <h3 className="font-semibold text-sm sm:text-base text-[#1A0033]">
-                    Update Password
-                  </h3>
-                  <div className="text-xs sm:text-sm text-gray-500 mt-1">
-                    We will send a secure reset link to your email.
-                  </div>
-                </div>
-                <button
-                  onClick={handleSendPasswordReset}
-                  disabled={isSendingPasswordReset}
-                  className="send-btn text-[#7C5CFC] text-sm sm:text-base font-semibold px-4 py-2 rounded-lg border border-[#d9d1fb] bg-white hover:bg-[#ede7fa] transition disabled:opacity-50 disabled:cursor-not-allowed"
-                >
-                  {isSendingPasswordReset ? "Sending..." : "Send Link"}
-                </button>
-              </div>
-
-              <div className="card delete-card flex flex-col sm:flex-row sm:items-center sm:justify-between bg-red-50 rounded-lg border border-red-200 p-4 sm:px-6 sm:py-4 gap-3 sm:gap-4">
-                <div className="flex-1 min-w-0">
-                  <h3 className="font-semibold text-sm sm:text-base text-[#E53E3E]">
-                    Delete Account
-                  </h3>
-                  <div className="text-xs sm:text-sm text-red-400 mt-1">
-                    Once you delete the account, your data cannot be retrieved. Be Certain!
-                  </div>
-                </div>
-                <button
-                  onClick={handleDeleteAccount}
-                  className="delete-btn text-[#E53E3E] text-xl sm:text-2xl focus:outline-none shrink-0"
-                >
-                  <Image
-                    src="/icons/icon2.png"
-                    alt="Chevron Down"
-                    width={24}
-                    height={24}
-                    className="feather feather-chevron-down"
-                  />
-                </button>
-              </div>
-            </div>
-          </div>
-        </div>
-      )}
+      {settings.activeTab === "Account" && <AccountTab settings={settings} />}
       <ToastContainer position="top-right" autoClose={3000} />
     </div>
   );
