+++ app/share/[slug]/ShareVideoPageClient.tsx
@@ -1,10 +1,13 @@
 "use client";
 
-import Image from "next/image";
-import Link from "next/link";
 import { useSession } from "next-auth/react";
-import { useEffect, useRef, useState } from "react";
 import type { CSSProperties } from "react";
+
+import { useViewTracking } from "./hooks/useViewTracking";
+import { getPreviewStage } from "./utils/previewStage";
+import ShareHeader from "./components/ShareHeader";
+import ShareSignupCTA from "./components/ShareSignupCTA";
+
 type ShareVideoPageClientProps = {
   title: string;
   description: string | null;
@@ -24,104 +27,17 @@ export default function ShareVideoPageClient({
   demoId,
   videoId,
 }: ShareVideoPageClientProps) {
-  const [viewId, setViewId] = useState<string | null>(null);
-  const videoRef = useRef<HTMLVideoElement>(null);
-  const watchedDurationRef = useRef<number>(0);
+  const videoRef = useViewTracking(demoId, videoId);
 
-  useEffect(() => {
-    // Record a view when the page loads
-    fetch("/api/views", {
-      method: "POST",
-      headers: { "Content-Type": "application/json" },
-      body: JSON.stringify({ demoId, exportedVideoId: videoId }),
-    })
-      .then((res) => res.json())
-      .then((data) => {
-        if (data.success && data.viewId && data.viewId !== "deduped") {
-          setViewId(data.viewId);
-        }
-      })
-      .catch(console.error);
-  }, [demoId, videoId]);
-
-  useEffect(() => {
-    if (!viewId) {
-      return;
-    }
-
-    // Periodically update the view duration on the server if the video is playing
-    const interval = setInterval(() => {
-      const vid = videoRef.current;
-      if (vid && !vid.paused) {
-        // Simple heuristic: just increment by the interval duration
-        // (A more robust way would be to track actual played ranges)
-        watchedDurationRef.current += 5;
-        fetch("/api/views", {
-          method: "POST",
-          headers: { "Content-Type": "application/json" },
-          body: JSON.stringify({
-            viewId,
-            duration: watchedDurationRef.current,
-          }),
-        }).catch(console.error);
-      }
-    }, 5000);
-
-    return () => clearInterval(interval);
-  }, [viewId]);
   const { status } = useSession();
   const isLoggedIn = status === "authenticated";
 
-  const ratioMap: Record<string, string> = {
-    "16:9": "16 / 9",
-    "1:1": "1 / 1",
-    "4:5": "4 / 5",
-    "9:16": "9 / 16",
-    "3:4": "3 / 4",
-  };
-
-  const previewFrameAspectRatio =
-    aspectRatio === "native" ? "16 / 9" : ratioMap[aspectRatio] || "16 / 9";
-  const [w, h] = previewFrameAspectRatio.split("/").map((v) => Number(v.trim()));
-  const previewRatioValue =
-    Number.isFinite(w) && Number.isFinite(h) && w > 0 && h > 0 ? w / h : 16 / 9;
-  const isPortraitPreview = previewRatioValue < 1;
-  const isSquarePreview = Math.abs(previewRatioValue - 1) < 0.05;
-
-  const stageWidth = isPortraitPreview
-    ? "min(420px, 52vw)"
-    : isSquarePreview
-      ? "min(700px, 74vw)"
-      : "min(1120px, 94vw)";
-  const stageHeight = isPortraitPreview ? "88%" : "84%";
-  const stageMaxWidth = isPortraitPreview ? "46%" : isSquarePreview ? "62%" : "92%";
+  const { previewFrameAspectRatio, stageWidth, stageHeight, stageMaxWidth } =
+    getPreviewStage(aspectRatio);
 
   return (
     <main className="min-h-screen bg-[#F2EDFF]">
-      <header className="sticky top-0 z-20 border-b border-[#E5DCFF] bg-white/80 backdrop-blur-md">
-        <div className="mx-auto flex h-16 w-full max-w-7xl items-center justify-between px-6">
-          <Link href="/" className="flex items-center gap-2 text-[#6B5ED8]">
-            <Image src="/icons/logo.png" alt="Marvedge logo" width={28} height={28} />
-            <span className="text-sm font-bold uppercase tracking-[0.18em]">Marvedge</span>
-          </Link>
-          {!isLoggedIn && (
-            <div className="flex items-center gap-3">
-              <Link
-                href="/auth/signin"
-                className="rounded-full border border-[#D4CAFF] px-4 py-2 text-sm font-medium text-[#6F5FBC]"
-              >
-                Login
-              </Link>
-              <Link
-                href="/auth/signup"
-                className="rounded-full bg-[#6E5AD8] px-4 py-2 text-sm font-semibold text-white"
-              >
-                Sign up
-              </Link>
-            </div>
-          )}
-        </div>
-      </header>
+      <ShareHeader isLoggedIn={isLoggedIn} />
 
       <section className="mx-auto w-full max-w-7xl px-4 pb-16 pt-8">
         <div className="mb-6 text-center">
@@ -163,28 +79,7 @@ export default function ShareVideoPageClient({
           </div>
         </div>
 
-        {!isLoggedIn && (
-          <div className="mt-10 text-center">
-            <h2 className="text-2xl font-semibold text-[#2E215D]">New to Marvedge?</h2>
-            <p className="mx-auto mt-2 max-w-2xl text-sm text-[#7E73AC]">
-              Start making your demos today and share polished walkthroughs in minutes.
-            </p>
-            <div className="mt-5 flex justify-center gap-3">
-              <Link
-                href="/auth/signup"
-                className="rounded-full bg-[#6A56D8] px-6 py-2.5 text-sm font-semibold text-white"
-              >
-                Sign up
-              </Link>
-              <Link
-                href="/auth/signin"
-                className="rounded-full border border-[#CFC2FF] bg-white px-6 py-2.5 text-sm font-semibold text-[#5D4BC5]"
-              >
-                Login
-              </Link>
-            </div>
-          </div>
-        )}
+        {!isLoggedIn && <ShareSignupCTA />}
       </section>
     </main>
   );
